Updated the sidebar for Outreach, Profile Manager
